api/requirements.txt
youtube-transcript-api
deepgram-sdk
yt-dlp
python-dotenv
